# src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Quiz API",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Home endpoint",
                "responses": {
                    "200": {
                        "description": "Returns a greeting message",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string",
                                    "example": "Hello Hono!"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/get-all-questions": {
            "get": {
                "summary": "Get all questions with options",
                "responses": {
                    "200": {
                        "description": "A list of questions with options",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "questions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/QuestionResponse"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload-questions": {
            "post": {
                "summary": "Upload a list of questions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "questions": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Question"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Questions uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Questions uploaded successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post-options": {
            "post": {
                "summary": "Post options for questions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "options": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Option"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Options posted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Options posted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post-answers": {
            "post": {
                "summary": "Post answers for questions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Answer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answers posted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Answers posted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/return-scores": {
            "post": {
                "summary": "Return scores based on answers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns the score of the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "score": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/delete-question": {
            "delete": {
                "summary": "Deletes a question",
                "operationId": "deleteQuestion",
                "tags": [
                    "Question"
                ],
                "parameters": [
                    {
                        "name": "questionID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "The ID of the question to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/delete-option": {
            "delete": {
                "summary": "Deletes an option",
                "operationId": "deleteOption",
                "tags": [
                    "Option"
                ],
                "parameters": [
                    {
                        "name": "optionID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "The ID of the option to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Option deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/update-question": {
            "patch": {
                "summary": "Updates a question",
                "operationId": "updateQuestion",
                "tags": [
                    "Question"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "questionID": {
                                        "type": "integer",
                                        "description": "The ID of the question to update"
                                    },
                                    "questionText": {
                                        "type": "string",
                                        "description": "The new text for the question"
                                    }
                                },
                                "required": [
                                    "questionID",
                                    "questionText"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Question updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "QuestionResponse": {
                "type": "object",
                "properties": {
                    "question": {
                        "type": "string"
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OptionID"
                        }
                    }
                }
            },
            "Question": {
                "type": "object",
                "properties": {
                    "QuestionText": {
                        "type": "string"
                    },
                    "QuestionID": {
                        "type": "integer"
                    }
                }
            },
            "Option": {
                "type": "object",
                "properties": {
                    "OptionID": {
                        "type": "integer"
                    },
                    "text": {
                        "type": "string"
                    },
                    "refQID": {
                        "type": "integer"
                    }
                }
            },
            "OptionID": {
                "type": "object",
                "properties": {
                    "OptionID": {
                        "type": "integer"
                    },
                    "text": {
                        "type": "string"
                    }
                }
            },
            "Answer": {
                "type": "object",
                "properties": {
                    "questionID": {
                        "type": "integer"
                    },
                    "optionsID": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                }
            }
        }
    }
}